+++ src/core/jsproxy.c
@@ -59,6 +59,8 @@
 #define IS_OBJECT_MAP (1 << 14)
 #define IS_ASYNC_ITERABLE (1 << 15)
 #define IS_GENERATOR (1 << 16)
+#define IS_ASYNC_GENERATOR (1 << 17)
+
 // clang-format on
 
 _Py_IDENTIFIER(get_event_loop);
@@ -471,11 +473,21 @@ static PyMethodDef JsGenerator_send_MethodDef = {
 
 static PyObject* Exc_JsException;
 
-static PyObject*
-JsGenerator_throw_inner(PyObject* self,
-                        PyObject* typ,
-                        PyObject* val,
-                        PyObject* tb)
+/**
+ * Shared logic between throw and async throw.
+ *
+ * Possibly "typ" is an exception instance and val and tb are null. Otherwise,
+ * it's an old style call "typ" should be an exception type, "val" an instance,
+ * and tb an optional traceback. Figure out which is the case and get an
+ * exception object.
+ *
+ * Then if the exception object is PyExc_GeneratorExit, call jsobj.return().
+ * Otherwise, convert it to js and call jsobj.throw(jsexc). Return the result of
+ * whichever of these two calls we make (or set the error flag and return NULL
+ * if something goes wrong).
+ */
+JsRef
+process_throw_args(PyObject* self, PyObject* typ, PyObject* val, PyObject* tb)
 {
   if (tb == Py_None) {
     tb = NULL;
@@ -521,11 +533,10 @@ JsGenerator_throw_inner(PyObject* self,
   }
 
   PyErr_Restore(typ, val, tb);
-  JsRef throw_res = NULL;
-  PyObject* result = NULL;
+  JsRef res = NULL;
   if (PyErr_ExceptionMatches(PyExc_GeneratorExit)) {
     PyErr_Clear();
-    throw_res = hiwire_CallMethodId_NoArgs(JsProxy_REF(self), &JsId_return);
+    res = hiwire_CallMethodId_NoArgs(JsProxy_REF(self), &JsId_return);
   } else {
     JsRef exc;
     if (PyErr_ExceptionMatches(Exc_JsException)) {
@@ -537,10 +548,30 @@ JsGenerator_throw_inner(PyObject* self,
     } else {
       exc = wrap_exception(); // cannot fail.
     }
-    throw_res = hiwire_CallMethodId_OneArg(JsProxy_REF(self), &JsId_throw, exc);
+    res = hiwire_CallMethodId_OneArg(JsProxy_REF(self), &JsId_throw, exc);
     hiwire_CLEAR(exc);
   }
+  return res;
+
+failed_throw:
+  /* Didn't use our arguments, so restore their original refcounts */
+  Py_DECREF(typ);
+  Py_XDECREF(val);
+  Py_XDECREF(tb);
+  return NULL;
+}
+
+static PyObject*
+JsGenerator_throw_inner(PyObject* self,
+                        PyObject* typ,
+                        PyObject* val,
+                        PyObject* tb)
+{
+  JsRef throw_res = NULL;
+  PyObject* result = NULL;
+  throw_res = process_throw_args(self, typ, val, tb);
   FAIL_IF_NULL(throw_res);
+  console_error_obj(throw_res);
   PySendResult ret = handle_next_result(throw_res, &result);
   if (ret == PYGEN_RETURN) {
     if (result == Py_None) {
@@ -553,13 +584,6 @@ JsGenerator_throw_inner(PyObject* self,
 finally:
   hiwire_CLEAR(throw_res);
   return result;
-
-failed_throw:
-  /* Didn't use our arguments, so restore their original refcounts */
-  Py_DECREF(typ);
-  Py_XDECREF(val);
-  Py_XDECREF(tb);
-  return NULL;
 }
 
 static PyObject*
@@ -631,73 +655,126 @@ JsProxy_GetAsyncIter(PyObject* self)
   return result;
 }
 
+/**
+ * The innermost layer of the result handling. This is called when the promise
+ * resolves. status tells us whether we are:
+ *  1. returning
+ *  0. yielding
+ * -1. raising
+ *
+ *
+ * jsvalue is either the result that we are returning/yielding or the error we
+ * are raising. closing tells us whether ``close`` was called, if so we raise
+ * an error if there is a yield in a finally block.
+ */
 void
-JsProxy_anext_js_stopiter(PyObject* set_exception, JsRef jsvalue)
+_agen_handle_result_js_c(PyObject* set_result,
+                         PyObject* set_exception,
+                         int status,
+                         JsRef jsvalue,
+                         bool closing)
 {
   PyObject* pyvalue = NULL;
   PyObject* e = NULL;
-  PyObject* result = NULL;
+
+  if (closing && status == 0) {
+    // If closing, status should not be yielding.
+    PyErr_SetString(PyExc_RuntimeError, "JavaScript generator ignored return");
+    goto return_error;
+  }
 
   pyvalue = js2python(jsvalue);
   if (pyvalue == NULL) {
-    // Uhhh... not sure what to do here... hopefully this won't happen
+    // hopefully this won't happen...
     PyErr_SetString(PyExc_RuntimeError, "Unable to get result");
+    goto return_error;
+  }
+
+  // TODO: generator lifetime handling
+  // if (status == 1 && pyproxy_Check(jsvalue)) {
+  //   destroy_proxy(jsvalue,
+  //                 "This borrowed proxy was automatically destroyed at the end
+  //                 " "of an async generator");
+  // }
+
+  if (status == 0) {
+    PyObject_CallOneArg(set_result, pyvalue);
+    // Not sure what to do if there is an error here...
+    goto finally;
+  }
+
+  if (status == 1) {
+    // Returning
+    e = PyObject_CallOneArg(PyExc_StopAsyncIteration, pyvalue);
+    // Not sure what to do if there is an error here...
+    goto return_error;
+  }
+
+  if (status == -1) {
+    Py_INCREF(pyvalue);
+    e = pyvalue;
+    goto return_error;
+  }
+
+return_error:
+  if (e == NULL) {
+    // Grab e from error flag
     PyObject *tp, *tb;
     PyErr_Fetch(&tp, &e, &tb);
     Py_CLEAR(tp);
     Py_CLEAR(tb);
-  } else {
-    e = PyObject_CallOneArg(PyExc_StopAsyncIteration, pyvalue);
   }
-  if (e != NULL) {
-    PyObject_CallOneArg(set_exception, e);
-  } else {
-    // We can't really handle an error here.
-    // Panic?
+  if (closing && (PyErr_GivenExceptionMatches(e, PyExc_StopAsyncIteration) ||
+                  PyErr_GivenExceptionMatches(e, PyExc_GeneratorExit))) {
+    PyObject_CallOneArg(set_result, Py_None);
+    goto finally;
   }
+
+  PyObject_CallOneArg(set_exception, e);
+  // Don't know what to do if there was an error...
+  PyErr_Clear();
+  goto finally;
+
+finally:
   Py_CLEAR(pyvalue);
   Py_CLEAR(e);
-  Py_CLEAR(result);
 }
 
 // clang-format off
 EM_JS_NUM(
 int,
-JsProxy_anext_js,
-(JsRef idobj, JsRef argid, char** msg, PyObject* set_result, PyObject* set_exception),
+_agen_handle_result_js,
+(JsRef promiseid, char** msg, PyObject* set_result, PyObject* set_exception, bool closing),
 {
-  let jsobj = Hiwire.get_value(idobj);
-  let jsarg;
-  if (argid !== 0) {
-    jsarg = Hiwire.get_value(argid);
-  }
-  let p = jsobj.next(jsarg);
-  let msg_ptr;
+  let p = Hiwire.get_value(promiseid);
+  // First check that p is a proper promise, if not return the error message for
+  // the type error.
+  let errmsg;
   if(typeof p !== "object") {
-    msg_ptr = stringToNewUTF8(`Result of anext() should be object not ${typeof p}`)
+    errmsg = `Result of anext() should be object not ${typeof p}`;
   } else if(typeof p.then !== "function") {
     if (typeof p.done === "boolean") {
-      msg_ptr = stringToNewUTF8(`Result of anext() was not a promise, use next() instead.`)
+      errmsg = `Result of anext() was not a promise, use next() instead.`;
     } else {
-      msg_ptr = stringToNewUTF8(`Result of anext() was not a promise.`)
+      errmsg = `Result of anext() was not a promise.`;
     }
   }
-  if (msg_ptr) {
-    DEREF_U32(msg, 0) = msg_ptr;
+  if (errmsg) {
+    DEREF_U32(msg, 0) = stringToNewUTF8(errmsg);
     return -1;
   }
+  // We need to hold onto set_result and set_exception until the promise resolves.
   _Py_IncRef(set_result);
   _Py_IncRef(set_exception);
+  // Call back into C to create the response.
   p.then(({done, value}) => {
-    if(!done) {
-      Module.callPyObject(set_result, [value]);
-      return;
-    }
     let id = Hiwire.new_value(value);
-    _JsProxy_anext_js_stopiter(set_exception, id);
+    __agen_handle_result_js_c(set_result, set_exception, done, id, closing);
     Hiwire.decref(id);
   }, (err) => {
-    Module.callPyObject(set_result, [err]);
+    let id = Hiwire.new_value(err);
+    __agen_handle_result_js_c(set_result, set_exception, -1, id, closing);
+    Hiwire.decref(id);
   }).finally(() => {
     _Py_DecRef(set_result);
     _Py_DecRef(set_exception);
@@ -706,15 +783,20 @@ JsProxy_anext_js,
 });
 // clang-format on
 
-static PyObject*
-JsProxy_asend(PyObject* self, PyObject* arg)
+/**
+ * Common logic between asend, athrow and aclose. Handle the resulting promise
+ * returned from next, throw, or return.
+ *
+ * Create a future and attach the promise to the future using the helper
+ * anext_js.
+ */
+PyObject*
+_agen_handle_result(JsRef promise, bool closing)
 {
   bool success = false;
-  JsRef promise = NULL;
   PyObject* loop = NULL;
   PyObject* set_result = NULL;
   PyObject* set_exception = NULL;
-  JsRef jsarg = NULL;
   PyObject* result = NULL;
 
   loop = PyObject_CallNoArgs(asyncio_get_event_loop);
@@ -728,13 +810,9 @@ JsProxy_asend(PyObject* self, PyObject* arg)
   set_exception = _PyObject_GetAttrId(result, &PyId_set_exception);
   FAIL_IF_NULL(set_exception);
 
-  if (arg != NULL) {
-    jsarg = python2js(arg);
-  }
-
   char* msg = NULL;
   int status =
-    JsProxy_anext_js(JsProxy_REF(self), jsarg, &msg, set_result, set_exception);
+    _agen_handle_result_js(promise, &msg, set_result, set_exception, closing);
   if (status == -1) {
     if (msg) {
       PyErr_SetString(PyExc_TypeError, msg);
@@ -748,25 +826,94 @@ finally:
   if (!success) {
     Py_CLEAR(result);
   }
-  Py_CLEAR(promise);
   Py_CLEAR(loop);
   Py_CLEAR(set_result);
   Py_CLEAR(set_exception);
-  hiwire_CLEAR(jsarg);
   return result;
 }
 
-static PyMethodDef JsProxy_asend_MethodDef = {
+static PyObject*
+JsGenerator_asend(PyObject* self, PyObject* arg)
+{
+  JsRef jsarg = Js_undefined;
+  JsRef next_res = NULL;
+  PyObject* result = NULL;
+  if (arg != NULL) {
+    jsarg = python2js(arg);
+    FAIL_IF_NULL(jsarg);
+  }
+
+  next_res = hiwire_CallMethodId_OneArg(JsProxy_REF(self), &JsId_next, jsarg);
+  FAIL_IF_NULL(next_res);
+  result = _agen_handle_result(next_res, false);
+
+finally:
+  hiwire_CLEAR(jsarg);
+  hiwire_CLEAR(next_res);
+  return result;
+}
+
+static PyMethodDef JsGenerator_asend_MethodDef = {
   "asend",
-  (PyCFunction)JsProxy_asend,
+  (PyCFunction)JsGenerator_asend,
   METH_O,
 };
 
 static PyObject*
-JsProxy_anext(PyObject* self)
+JsGenerator_anext(PyObject* self)
 {
-  return JsProxy_asend(self, NULL);
+  return JsGenerator_asend(self, NULL);
 }
+
+static PyObject*
+JsGenerator_athrow(PyObject* self, PyObject* const* args, Py_ssize_t nargs)
+{
+  PyObject* typ;
+  PyObject* val = NULL;
+  PyObject* tb = NULL;
+
+  if (!_PyArg_ParseStack(args, nargs, "O|OO:athrow", &typ, &val, &tb)) {
+    return NULL;
+  }
+
+  JsRef throw_res = NULL;
+  PyObject* result = NULL;
+
+  throw_res = process_throw_args(self, typ, val, tb);
+  FAIL_IF_NULL(throw_res);
+  result = _agen_handle_result(throw_res, false);
+
+finally:
+  hiwire_CLEAR(throw_res);
+  return result;
+}
+
+static PyMethodDef JsGenerator_athrow_MethodDef = {
+  "athrow",
+  (PyCFunction)JsGenerator_athrow,
+  METH_FASTCALL,
+};
+
+static PyObject*
+JsGenerator_aclose(PyObject* self, PyObject* ignored)
+{
+  JsRef throw_res = NULL;
+  PyObject* result = NULL;
+
+  throw_res = process_throw_args(self, PyExc_GeneratorExit, NULL, NULL);
+  FAIL_IF_NULL(throw_res);
+  result = _agen_handle_result(throw_res, true);
+
+finally:
+  hiwire_CLEAR(throw_res);
+  return result;
+}
+
+static PyMethodDef JsGenerator_aclose_MethodDef = {
+  "aclose",
+  (PyCFunction)JsGenerator_aclose,
+  METH_NOARGS,
+};
 
 /**
  * This is exposed as a METH_NOARGS method on the JsProxy. It returns
@@ -3329,10 +3476,10 @@ JsProxy_create_subtype(int flags)
     slots[cur_slot++] =
       (PyType_Slot){ .slot = Py_am_aiter, .pfunc = (void*)PyObject_SelfIter };
     slots[cur_slot++] =
-      (PyType_Slot){ .slot = Py_am_anext, .pfunc = (void*)JsProxy_anext };
+      (PyType_Slot){ .slot = Py_am_anext, .pfunc = (void*)JsGenerator_anext };
     // Send works okay on any js object that has a "next" method
     methods[cur_method++] = JsGenerator_send_MethodDef;
-    methods[cur_method++] = JsProxy_asend_MethodDef;
+    methods[cur_method++] = JsGenerator_asend_MethodDef;
   }
   if (flags & IS_GENERATOR) {
     // throw and close need "throw" and "return" methods to work. We currently
@@ -3341,6 +3488,15 @@ JsProxy_create_subtype(int flags)
     // to Generator.
     methods[cur_method++] = JsGenerator_throw_MethodDef;
     methods[cur_method++] = JsGenerator_close_MethodDef;
+  }
+
+  if (flags & IS_ASYNC_GENERATOR) {
+    // throw and close need "throw" and "return" methods to work. We currently
+    // don't trust that an object with "next", "throw", and "return" is a
+    // generator though -- we require that it actually have it's toStringTag set
+    // to Generator.
+    methods[cur_method++] = JsGenerator_athrow_MethodDef;
+    methods[cur_method++] = JsGenerator_aclose_MethodDef;
   }
 
   if (flags & HAS_LENGTH) {
@@ -3595,6 +3751,7 @@ EM_JS_NUM(int, compute_typeflags, (JsRef idobj), {
   SET_FLAG_IF(IS_TYPEDARRAY,
               ArrayBuffer.isView(obj) && obj.constructor.name !== "DataView");
   SET_FLAG_IF(IS_GENERATOR, typeTag === "[object Generator]");
+  SET_FLAG_IF(IS_ASYNC_GENERATOR, typeTag === "[object AsyncGenerator]");
   // clang-format on
   return type_flags;
 });
@@ -3839,6 +3996,7 @@ JsProxy_init(PyObject* core_module)
   AddFlag(IS_OBJECT_MAP);
   AddFlag(IS_ASYNC_ITERABLE);
   AddFlag(IS_GENERATOR);
+  AddFlag(IS_ASYNC_GENERATOR);
 
 #undef AddFlag